requirements.txt
openai
requests
python-dotenv
weaviate-client